Refactor IProducer namespace
