+++ src/EsbcProducer/Repositories/IProducer.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 
-namespace EsbcProducer.Services
+namespace EsbcProducer.Repositories
 {
     public interface IProducer
     {
