+++ src/EsbcProducer/Worker/Impl/Proletarian.cs
@@ -1,4 +1,4 @@
-﻿using EsbcProducer.Services;
+﻿using EsbcProducer.Repositories;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
